Implemented more tests

+++ test/stages/buildAndPushDocker.test.ts
@@ -29,30 +29,37 @@ const prepareEnvMock = () =>
 		NEXUS_DOCKER_PASSWORD: 'password'
 	}));
 
-const validateCommands = () => {
-	expect(runCommandMock).toHaveBeenCalledTimes(5);
+const validateCommands = (numCommands = 5) => {
+	expect(runCommandMock).toHaveBeenCalledTimes(numCommands);
+	let callCount = 1;
 	expect(runCommandMock).toHaveBeenNthCalledWith(
-		1,
+		callCount,
 		'sudo docker login craigmiller160.ddns.net:30004 -u ${user} -p ${password}',
 		{ printOutput: true, variables: { user: 'user', password: 'password' } }
 	);
+	callCount++;
 	expect(runCommandMock).toHaveBeenNthCalledWith(
-		2,
+		callCount,
 		'sudo docker image ls | grep my-project | grep 1.0.0 || true',
 		{ printOutput: true }
 	);
+	callCount++;
+	if (numCommands === 5) {
+		expect(runCommandMock).toHaveBeenNthCalledWith(
+			3,
+			"sudo docker image ls | grep my-project | grep 1.0.0 | awk '{ print $3 }' | xargs sudo docker image rm -f",
+			{ printOutput: true }
+		);
+		callCount++;
+	}
 	expect(runCommandMock).toHaveBeenNthCalledWith(
-		3,
-		"sudo docker image ls | grep my-project | grep 1.0.0 | awk '{ print $3 }' | xargs sudo docker image rm -f",
-		{ printOutput: true }
-	);
-	expect(runCommandMock).toHaveBeenNthCalledWith(
-		4,
+		callCount,
 		'sudo docker build --network=host -t craigmiller160.ddns.net:30004/my-project:1.0.0',
 		{ printOutput: true }
 	);
+	callCount++;
 	expect(runCommandMock).toHaveBeenNthCalledWith(
-		5,
+		callCount,
 		'sudo docker push craigmiller160.ddns.net:30004/my-project:1.0.0',
 		{ printOutput: true }
 	);
@@ -113,7 +120,19 @@ describe('buildAndPushDocker', () => {
 	});
 
 	it('builds and pushes docker image for maven application, with no existing images', async () => {
-		throw new Error();
+		runCommandMock.mockReset();
+		runCommandMock.mockImplementation(() => TE.right(''));
+		prepareEnvMock();
+
+		const buildContext: BuildContext = {
+			...baseBuildContext,
+			projectType: ProjectType.MavenApplication
+		};
+
+		const result = await buildAndPushDocker.execute(buildContext)();
+		expect(result).toEqualRight(buildContext);
+
+		validateCommands(4);
 	});
 
 	it('builds and pushes docker image for npm application', async () => {
